Allow shared layer VMLockingPrimitive in extension layers
