+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/locks/VMLockFeature.java
@@ -1,0 +1,60 @@
+/*
+ * Copyright (c) 2026, 2026, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package com.oracle.svm.hosted.locks;
+
+import org.graalvm.nativeimage.ImageSingletons;
+
+import com.oracle.graal.pointsto.heap.ImageHeapScanner;
+import com.oracle.svm.core.feature.InternalFeature;
+import com.oracle.svm.core.imagelayer.ImageLayerBuildingSupport;
+import com.oracle.svm.core.locks.VMLockSupport;
+import com.oracle.svm.core.locks.VMLockingPrimitive;
+import com.oracle.svm.hosted.FeatureImpl.DuringSetupAccessImpl;
+import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
+import com.oracle.svm.shared.util.VMError;
+
+@AutomaticallyRegisteredFeature
+final class VMLockFeature implements InternalFeature {
+    @Override
+    public void afterRegistration(AfterRegistrationAccess access) {
+        if (ImageLayerBuildingSupport.firstImageBuild()) {
+            ImageSingletons.add(VMLockSupport.class, new VMLockSupport());
+        }
+    }
+
+    @Override
+    public void duringSetup(DuringSetupAccess access) {
+        if (ImageLayerBuildingSupport.firstImageBuild()) {
+            /* Replace build-time with run-time objects. */
+            VMLockSupport support = ImageSingletons.lookup(VMLockSupport.class);
+            support.registerObjectReplacers(access);
+        } else {
+            /* Extension layers must not add any extra VMLockingPrimitives. */
+            ImageHeapScanner imageHeapScanner = ((DuringSetupAccessImpl) access).getBigBang().getUniverse().getHeapScanner();
+            access.registerObjectReachabilityHandler((obj) -> VMError.guarantee(imageHeapScanner.isObjectInSharedLayer(obj),
+                            "A VMLockingPrimitive is added in an extension layer, which is unsupported: %s", obj), VMLockingPrimitive.class);
+        }
+    }
+}
